consolidate mTLS helper into manifestutils.IsTAMTLSEnabled
