+++ internal/manifests/manifestutils/mtls_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package targetallocator
+package manifestutils
 
 import (
 	"testing"
@@ -14,18 +14,18 @@ import (
 	"github.com/open-telemetry/opentelemetry-operator/internal/config"
 )
 
-func TestIsMTLSEnabledDefaultsUseCertManagerToTrue(t *testing.T) {
-	ta := v1alpha1.TargetAllocator{}
+func TestIsTAMTLSEnabledDefaultsUseCertManagerToTrue(t *testing.T) {
+	ta := &v1alpha1.TargetAllocator{}
 	ta.Spec.Mtls = &v1beta1.TargetAllocatorMTLS{Enabled: true}
 
 	cfg := config.Config{CertManagerAvailability: certmanager.Available}
 
-	assert.True(t, isMTLSEnabled(cfg, ta))
+	assert.True(t, IsTAMTLSEnabled(cfg, ta))
 }
 
-func TestIsMTLSEnabledDisabledWhenUseCertManagerFalse(t *testing.T) {
+func TestIsTAMTLSEnabledDisabledWhenUseCertManagerFalse(t *testing.T) {
 	useCertManager := false
-	ta := v1alpha1.TargetAllocator{}
+	ta := &v1alpha1.TargetAllocator{}
 	ta.Spec.Mtls = &v1beta1.TargetAllocatorMTLS{
 		Enabled:        true,
 		UseCertManager: &useCertManager,
@@ -33,31 +33,37 @@ func TestIsMTLSEnabledDisabledWhenUseCertManagerFalse(t *testing.T) {
 
 	cfg := config.Config{CertManagerAvailability: certmanager.Available}
 
-	assert.False(t, isMTLSEnabled(cfg, ta))
+	assert.False(t, IsTAMTLSEnabled(cfg, ta))
 }
 
-func TestIsMTLSEnabledNilMtls(t *testing.T) {
-	ta := v1alpha1.TargetAllocator{}
+func TestIsTAMTLSEnabledNilTA(t *testing.T) {
+	cfg := config.Config{CertManagerAvailability: certmanager.Available}
+
+	assert.False(t, IsTAMTLSEnabled(cfg, nil))
+}
+
+func TestIsTAMTLSEnabledNilMtls(t *testing.T) {
+	ta := &v1alpha1.TargetAllocator{}
 
 	cfg := config.Config{CertManagerAvailability: certmanager.Available}
 
-	assert.False(t, isMTLSEnabled(cfg, ta))
+	assert.False(t, IsTAMTLSEnabled(cfg, ta))
 }
 
-func TestIsMTLSEnabledDisabledMtls(t *testing.T) {
-	ta := v1alpha1.TargetAllocator{}
+func TestIsTAMTLSEnabledDisabledMtls(t *testing.T) {
+	ta := &v1alpha1.TargetAllocator{}
 	ta.Spec.Mtls = &v1beta1.TargetAllocatorMTLS{Enabled: false}
 
 	cfg := config.Config{CertManagerAvailability: certmanager.Available}
 
-	assert.False(t, isMTLSEnabled(cfg, ta))
+	assert.False(t, IsTAMTLSEnabled(cfg, ta))
 }
 
-func TestIsMTLSEnabledCertManagerUnavailable(t *testing.T) {
-	ta := v1alpha1.TargetAllocator{}
+func TestIsTAMTLSEnabledCertManagerUnavailable(t *testing.T) {
+	ta := &v1alpha1.TargetAllocator{}
 	ta.Spec.Mtls = &v1beta1.TargetAllocatorMTLS{Enabled: true}
 
 	cfg := config.Config{CertManagerAvailability: certmanager.NotAvailable}
 
-	assert.False(t, isMTLSEnabled(cfg, ta))
+	assert.False(t, IsTAMTLSEnabled(cfg, ta))
 }
